changed core files from tsx to ts and updated links
