+++ src/containers/forgotPassword/ResetPassword.tsx
@@ -7,9 +7,8 @@ import { connect } from "react-redux";
 import { Grid } from "@material-ui/core";
 
 import { useDispatch } from "react-redux";
-import { SetAuthenticated } from "../../core/store/app/appActions";
+import { SetAuthenticated } from "../../core/store/app/actions";
 
-// import { useDispatch } from "react-redux";
 import ResetPasswordForm from "../../components/forgotPassword/ResetPasswordForm";
 const mainImage = require("../../assets/images/main-image.jpg");
