+++ src/containers/register/Register.tsx
@@ -3,12 +3,11 @@ import { Typography } from "@material-ui/core";
 import { APP_NAME } from "../../core/constants";
 
 import { connect } from "react-redux";
-// import "./login.css";
 
 import { Grid } from "@material-ui/core";
 
 import { useDispatch } from "react-redux";
-import { SetAuthenticated } from "../../core/store/app/appActions";
+import { SetAuthenticated } from "../../core/store/app/actions";
 import RegisterForm from "../../components/register/RegisterForm";
 const mainImage = require("../../assets/images/main-image.jpg");
