+++ src/core/services/privacypolicy.service.ts
@@ -1,9 +1,5 @@
 import axios from "axios";
 
-import { useDispatch } from "react-redux";
-
-// import { acceptPrivacyPolicy as acceptPolicy } from "../store/auth/authActions";
-
 export const acceptPrivacyPolicy = (accessToken: string) => {
   const jwtToken = `Bearer ${accessToken}`;
   axios
@@ -18,7 +14,6 @@ export const acceptPrivacyPolicy = (accessToken: string) => {
     )
     .then((response) => {
       console.log("Response", response.data);
-      //   dispatch(LoginUser(username, password, rememberMe));
     })
     .catch((err) => {
       console.log("error", err);
