+++ src/core/store/app/reducer.ts
@@ -1,4 +1,4 @@
-import { SET_AUTHENTICATED } from "./appTypes";
+import { SET_AUTHENTICATED } from "./types";
 
 interface DefaultStateI {}
 
@@ -8,8 +8,6 @@ export const appReducer: any = (
   state: DefaultStateI = defaultState,
   action: any
 ) => {
-  //   const { LOGIN_USER } = CONSTANTS;
-  console.log("inside app reducers");
   switch (action.type) {
     case SET_AUTHENTICATED:
       return action.payload;
