+++ src/core/store/types.ts
@@ -2,15 +2,3 @@ export const CONSTANTS = {
   SET_AUTHENTICATED: "SET_AUTHENTICATED",
   LOGIN_USER: "LOGIN_USER",
 };
-
-// TODO: Add types for login and register
-
-// //user reducer types
-// export const SET_UNAUTHENTICATED=’SET_UNAUTHENTICATED’;
-// export const SET_USER=’SET_USER’;
-// export const LOADING_USER=’LOADING_USER’;
-
-// //UI reducer types
-// export const SET_ERRORS=’SET_ERRORS’;
-// export const LOADING_UI=’LOADING_UI’;
-// export const CLEAR_ERRORS=’CLEAR_ERRORS’;
